admin-a/backend/app/pyproject.toml
[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = "A simple FastAPI application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
fastapi = "^0.109.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptgraphy"], version = "^3.3.0"}
pytest = "^8.3.1"
httpx = "^0.26.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
requests = "^2.32.3"
boto3 = "^1.34.145"

[tool.poetry.dev-dependencies]
yapf = "^0.3.1"
isort = "^5.13.2"

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

